requirements.txt
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
cryptography==44.0.0
decorator==5.1.1
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pycparser==2.22
python-telegram-bot==21.10
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.12.2
